# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse parts management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateCategoryInput {
  name: string;
  description?: string | null;
}

export interface Supplier {
  id: number;
  name: string;
  contactPerson?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateSupplierInput {
  name: string;
  contactPerson?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  notes?: string | null;
}

export interface Warehouse {
  id: number;
  name: string;
  location?: string | null;
  description?: string | null;
  createdAt: string;
}

export interface CreateWarehouseInput {
  name: string;
  location?: string | null;
  description?: string | null;
}

export interface Part {
  id: number;
  partNumber: string;
  name: string;
  description?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  supplierId?: number | null;
  supplierName?: string | null;
  warehouseId?: number | null;
  warehouseName?: string | null;
  quantity: number;
  minQuantity: number;
  unitPrice?: number | null;
  unit: string;
  compatibleModels?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePartInput {
  partNumber: string;
  name: string;
  description?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  warehouseId?: number | null;
  quantity: number;
  minQuantity: number;
  unitPrice?: number | null;
  unit: string;
  compatibleModels?: string | null;
  notes?: string | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  receipt: "receipt",
  issue: "issue",
  adjustment: "adjustment",
  transfer: "transfer",
} as const;

export interface Transaction {
  id: number;
  partId: number;
  partName?: string | null;
  partNumber?: string | null;
  type: TransactionType;
  quantity: number;
  previousQuantity: number;
  newQuantity: number;
  unitPrice?: number | null;
  reference?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateTransactionInputType =
  (typeof CreateTransactionInputType)[keyof typeof CreateTransactionInputType];

export const CreateTransactionInputType = {
  receipt: "receipt",
  issue: "issue",
  adjustment: "adjustment",
  transfer: "transfer",
} as const;

export interface CreateTransactionInput {
  partId: number;
  type: CreateTransactionInputType;
  quantity: number;
  unitPrice?: number | null;
  reference?: string | null;
  notes?: string | null;
}

export type DashboardStatsTopCategoriesItem = {
  name: string;
  count: number;
};

export interface DashboardStats {
  totalParts: number;
  totalItems: number;
  lowStockCount: number;
  outOfStockCount: number;
  totalValue: number;
  recentTransactions: Transaction[];
  topCategories: DashboardStatsTopCategoriesItem[];
}

export type ListPartsParams = {
  search?: string;
  categoryId?: number;
  supplierId?: number;
  lowStock?: boolean;
};

export type ListTransactionsParams = {
  partId?: number;
  type?: ListTransactionsType;
  limit?: number;
  offset?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  receipt: "receipt",
  issue: "issue",
  adjustment: "adjustment",
  transfer: "transfer",
} as const;