# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Warehouse parts management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: parts
    description: Parts catalog
  - name: suppliers
    description: Suppliers management
  - name: warehouses
    description: Warehouse locations
  - name: transactions
    description: Inventory transactions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /parts:
    get:
      operationId: listParts
      tags: [parts]
      summary: List all parts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Part"
    post:
      operationId: createPart
      tags: [parts]
      summary: Create a new part
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartInput"
      responses:
        "201":
          description: Created part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"

  /parts/{id}:
    get:
      operationId: getPart
      tags: [parts]
      summary: Get part by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Part details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePart
      tags: [parts]
      summary: Update a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartInput"
      responses:
        "200":
          description: Updated part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
    delete:
      operationId: deletePart
      tags: [parts]
      summary: Delete a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [parts]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [parts]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List all warehouse locations
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create a warehouse location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseInput"
      responses:
        "201":
          description: Created warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List inventory transactions
      parameters:
        - name: partId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [receipt, issue, adjustment, transfer]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create an inventory transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [parts]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt

    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required:
        - name

    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt

    CreateSupplierInput:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name

    Warehouse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt

    CreateWarehouseInput:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - name

    Part:
      type: object
      properties:
        id:
          type: integer
        partNumber:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        warehouseId:
          type: integer
          nullable: true
        warehouseName:
          type: string
          nullable: true
        quantity:
          type: integer
        minQuantity:
          type: integer
        unitPrice:
          type: number
          nullable: true
        unit:
          type: string
        compatibleModels:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - partNumber
        - name
        - quantity
        - minQuantity
        - unit
        - createdAt
        - updatedAt

    CreatePartInput:
      type: object
      properties:
        partNumber:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        warehouseId:
          type: integer
          nullable: true
        quantity:
          type: integer
        minQuantity:
          type: integer
        unitPrice:
          type: number
          nullable: true
        unit:
          type: string
        compatibleModels:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - partNumber
        - name
        - quantity
        - minQuantity
        - unit

    Transaction:
      type: object
      properties:
        id:
          type: integer
        partId:
          type: integer
        partName:
          type: string
          nullable: true
        partNumber:
          type: string
          nullable: true
        type:
          type: string
          enum: [receipt, issue, adjustment, transfer]
        quantity:
          type: integer
        previousQuantity:
          type: integer
        newQuantity:
          type: integer
        unitPrice:
          type: number
          nullable: true
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - partId
        - type
        - quantity
        - previousQuantity
        - newQuantity
        - createdAt

    CreateTransactionInput:
      type: object
      properties:
        partId:
          type: integer
        type:
          type: string
          enum: [receipt, issue, adjustment, transfer]
        quantity:
          type: integer
        unitPrice:
          type: number
          nullable: true
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - partId
        - type
        - quantity

    DashboardStats:
      type: object
      properties:
        totalParts:
          type: integer
        totalItems:
          type: integer
        lowStockCount:
          type: integer
        outOfStockCount:
          type: integer
        totalValue:
          type: number
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        topCategories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
            required:
              - name
              - count
      required:
        - totalParts
        - totalItems
        - lowStockCount
        - outOfStockCount
        - totalValue
        - recentTransactions
        - topCategories