# lib/api-zod/src/generated/types/createPartInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse parts management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePartInput {
  partNumber: string;
  name: string;
  description?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  warehouseId?: number | null;
  quantity: number;
  minQuantity: number;
  unitPrice?: number | null;
  unit: string;
  compatibleModels?: string | null;
  notes?: string | null;
}