# lib/api-zod/src/generated/types/createSupplierInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse parts management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSupplierInput {
  name: string;
  contactPerson?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  notes?: string | null;
}