# lib/api-zod/src/generated/types/createTransactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse parts management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionInputType } from "./createTransactionInputType";

export interface CreateTransactionInput {
  partId: number;
  type: CreateTransactionInputType;
  quantity: number;
  unitPrice?: number | null;
  reference?: string | null;
  notes?: string | null;
}