# lib/api-zod/src/generated/types/warehouse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse parts management API
 * OpenAPI spec version: 0.1.0
 */

export interface Warehouse {
  id: number;
  name: string;
  location?: string | null;
  description?: string | null;
  createdAt: Date;
}